requirements.txt
opencv-python-headless
streamlit
numpy
pillow
